Finish likes and split post and feed service/controller
